# nlmeans_core: replace SIMD_FOR with for to improve performance