CPU: Intel(R) Core(TM) i7-6700
Threads: 4
Build: Release, SSE2-optimized codepaths disabled

Test: `0020-denoise-nlmeans`

**pixel pipeline** processing took:
`SIMD_FOR`: 7.9 sec (7.927; 7.889; 7.897)
`for`: 5.3 sec (5.285; 5.337; 5.272)

**CPU time** for [Loop at line 473 in nlmeans_denoise._omp_fn.0]
_Measured by Intel VTune Profiler_
`SIMD_FOR`: 19.010s
`for`: 7.424s